app/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --no-index --output-file requirements.txt requirements.in
#
daemonize==2.4.7
psutil==5.2.2
py==1.4.34                # via pytest
pytest==3.2.1
pyyaml==3.12
